requirements.txt
fastapi
uvicorn
streamlit
requests
docker
python-dotenv
transformers
torch